+++ app/frontend/src/javascript/components/payment/payzen/payzen-modal.tsx
@@ -25,7 +25,7 @@ interface PayZenModalProps {
  * This component enables the user to input his card data or process payments, using the PayZen gateway.
  * Supports Strong-Customer Authentication (SCA).
  *
- * This component should not be called directly. Prefer using <PaymentModal> which can handle the configuration
+ * This component should not be called directly. Prefer using <CardPaymentModal> which can handle the configuration
  *  of a different payment gateway.
  */
 export const PayZenModal: React.FC<PayZenModalProps> = ({ isOpen, toggleModal, afterSuccess, onError, cart, currentUser, schedule, customer }) => {