+++ app/services/prepaid_pack_service.rb
@@ -24,7 +24,7 @@ class PrepaidPackService
         .includes(:prepaid_pack)
         .references(:prepaid_packs)
         .where('statistic_profile_id = ?', user.statistic_profile.id)
-        .where('expires_at > ?', DateTime.current)
+        .where('expires_at > ? OR expires_at IS NULL', DateTime.current)
         .where('prepaid_packs.priceable_id = ?', priceable.id)
         .where('prepaid_packs.priceable_type = ?', priceable.class.name)
         .where('minutes_used < prepaid_packs.minutes')