+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2022_01_11_134253) do
+ActiveRecord::Schema.define(version: 2022_01_18_123741) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "fuzzystrmatch"